% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBox.R
\name{checkBox}
\alias{checkBox}
\title{Checkbox Question}
\usage{
checkBox(questionNum, questionBody, answerA, answerB, answerC, answerD)
}
\arguments{
\item{questionNum}{Integer, the question number}

\item{questionBody}{String, the question being asked}

\item{answerA}{String, option A}

\item{answerB}{String, option B}

\item{answerC}{String, option C}

\item{answerD}{String, option D}
}
\description{
Creates the UI for a multiple choice question using Checkboxes, allowing multiple answers to be selected.
}
\examples{
checkBox(1, "Sample Checkboxes", "Answer 1", "Answer 2", "Answer 3", "Answer 4")


}
