% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid-data.R
\docType{data}
\name{covid}
\alias{covid}
\title{San Diego Covid-19 data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(covid)
}
\description{
Data related to infections of covid-19 for each zip code in San Diego County as of October 1, 2020.
}
\references{
(2020) City of San Diego
}
\keyword{datasets}
