% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillBlank.R
\name{fillBlank}
\alias{fillBlank}
\title{Fill in the Blank Question}
\usage{
fillBlank(questionNum, questionBody, questionCode)
}
\arguments{
\item{questionNum}{Integer, the question number}

\item{questionBody}{String, the question being asked}
}
\description{
Creates the UI for a fill-in-the-blank question using textboxes, allowing a user input to be obtained.
}
