% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire-data.R
\docType{data}
\name{fire}
\alias{fire}
\title{San Diego Fire Incidents data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
https://data.sandiego.gov/datasets/fire-incidents/
}
\usage{
data(fire)
}
\description{
Fire and EMS calls dispatched by the San Diego Fire Communications Center.
}
\references{
(2020) City of San Diego
}
\keyword{datasets}
