% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeForm.R
\name{freeForm}
\alias{freeForm}
\title{Free Form Question}
\usage{
freeForm(questionNum, questionBody)
}
\arguments{
\item{questionNum}{Integer, the question number}

\item{questionBody}{String, the question being asked}
}
\description{
Creates the UI for a free form question using wide textboxes, allowing a user input to be obtained.
}
\examples{
freeForm(1, 'Provide code that evaluates to 5.')


}
