% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleList.R
\docType{data}
\name{getHist}
\alias{getHist}
\title{Example List}
\format{
An object of class \code{list} of length 3.
}
\usage{
getHist
}
\description{
Holds variables for use in building examples
}
\keyword{datasets}
