% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{'Hello' example function}
\usage{
hello(n)
}
\arguments{
\item{n}{integer; how many times 'Hello, World!' will be printed}
}
\description{
Function for saying hello multiple times.
}
\examples{
hello(n = 2)

}
