% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textList.R
\docType{data}
\name{introText}
\alias{introText}
\title{Text List}
\format{
An object of class \code{character} of length 1.
}
\usage{
introText
}
\description{
Holds variables for use in providing text to the Shiny app
}
\keyword{datasets}
