% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionList.R
\docType{data}
\name{q1Question}
\alias{q1Question}
\title{Question List}
\format{
An object of class \code{list} of length 6.
}
\usage{
q1Question
}
\description{
Holds variables for use in building questions
}
\keyword{datasets}
