% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\docType{data}
\name{questions}
\alias{questions}
\title{Holds variables pertaining to the questions asked in the app}
\format{
An object of class \code{character} of length 4.
}
\usage{
questions
}
\description{
Holds variables pertaining to the questions asked in the app
}
\keyword{datasets}
