% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionList.R
\docType{data}
\name{sampleQuestion}
\alias{sampleQuestion}
\title{Question List}
\format{
An object of class \code{list} of length 6.
}
\usage{
sampleQuestion
}
\description{
Holds variables for use in building questions
}
\keyword{datasets}
