% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny app UI}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
ui
}
\description{
Builds the visible content for the Shiny app to function.
}
\keyword{datasets}
