% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillBlank.R
\name{fillBlank}
\alias{fillBlank}
\title{Fill in the Blank Question}
\usage{
fillBlank(questionNum, questionBody)
}
\arguments{
\item{questionNum}{Integer, the question number}

\item{questionBody}{String, the question being asked}
}
\description{
Creates the UI for a fill-in-the-blank question using textboxes, allowing a user input to be obtained.
}
\examples{
fillBlank(1, "get_example <- function(x){\n  z <- x + ___\n  return(z)\n}\n> get_example(2)\''nOUTPUT: 3")


}
