% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillBlank.R
\name{fillBlank}
\alias{fillBlank}
\title{Fill in the Blank Example}
\usage{
fillBlank(exampleNum, exampleBody, exampleCode)
}
\arguments{
\item{exampleNum}{Integer, the example number}

\item{exampleBody}{String, the example question being asked}

\item{exampleCode}{String, the code provided to fill the blank in}
}
\description{
Creates the UI for a fill-in-the-blank example using textboxes, allowing a user input to be obtained.
}
