% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Placeholder text}
\format{
An object of class \code{data.frame} with 23509 rows and 3 columns.
}
\usage{
fire
}
\description{
Placeholder text
}
\keyword{datasets}
