% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multChoice.R
\name{multChoice}
\alias{multChoice}
\title{Multiple Choice Question}
\usage{
multChoice(questionNum, questionBody, answerA, answerB, answerC, answerD)
}
\arguments{
\item{questionNum}{Integer, the question number}

\item{questionBody}{String, the question being asked}

\item{answerA}{String, option A}

\item{answerB}{String, option B}

\item{answerC}{String, option C}

\item{answerD}{String, option D}
}
\description{
Creates the UI for a multiple choice question using Radio Buttons, allowing one answer to be selected.
}
\examples{
multChoice(1, "Sample Multiple Choice", "Answer 1", "Answer 2", "Answer 3", "Answer 4")


}
