% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny app Server}
\usage{
server(input, output)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
Obtains info and returns data to the UI for the Shiny app to function.
}
