% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{startApp}
\alias{startApp}
\title{Start Shiny App}
\usage{
startApp()
}
\value{
shiny application object
}
\description{
Launches the created shiny app
}
\examples{
\dontrun{startApp()}


}
